FederatedLearning-main/requirements.txt
google-api-core==2.3.0
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-cloud-pubsub
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
matplotlib==3.3.4
matplotlib-inline==0.1.3
numpy   
torch
torchvision==0.11.1